Refactor container and user interface
