--- S2VX.Game/Editor/UserInterface/TapPanel.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace S2VX.Game.Editor.UserInterface {
-    public class TapPanel : S2VXOverlayContainer {
-    }
-}
